support for building pal

+++ justfile
@@ -16,17 +16,21 @@ clean_all:
 # run this to setup your local machine for development
 setup:
     docker build -t {{ image }} .
-    docker run --rm -v $(pwd):/home/dev {{ image }} ./scripts/extract_baserom.u.sh
+    [ -f baserom.u.z64 ] && docker run --rm -v $(pwd):/home/dev {{ image }} ./scripts/extract_baserom.u.sh
+    [ -f baserom.e.z64 ] && docker run --rm -v $(pwd):/home/dev {{ image }} ./scripts/extract_diff.e.sh
 
 # build the rom
 make:
     if test -z "$(docker images -q {{ image }})"; then just setup; fi
     docker run --rm -v "$(pwd):/home/dev" {{ image }} make -j$(nproc) COMPARE=0 FINAL=YES
+    [ -f baserom.e.z64 ] && docker run --rm -v "$(pwd):/home/dev" {{ image }} make -j$(nproc) COMPARE=0 FINAL=YES VERSION=EU
 
 make-debug:
     docker run --rm -v "$(pwd):/home/dev" {{ image }} make -j$(nproc) COMPARE=0 FINAL=NO
+    [ -f baserom.e.z64 ] && docker run --rm -v "$(pwd):/home/dev" {{ image }} make -j$(nproc) COMPARE=0 FINAL=NO VERSION=EU
 
 # builds the rom and uploads it to a connected summercart64
 sc64: make
     sc64deployer sd mkdir /CUSTOM || true
-    sc64deployer sd upload build/u/ge007.u.z64 /CUSTOM/practice-rom_latest.z64
+    sc64deployer sd upload build/u/ge007.u.z64 /CUSTOM/practice-rom_latest_ntsc_u.z64
+    sc64deployer sd upload build/u/ge007.u.z64 /CUSTOM/practice-rom_latest_pal.z64
